# docs/api-reference/openapi.yaml
openapi: 3.1.0
...
info:
  title: DolphinFlashcardApp API
  version: '1.0'
  summary: Manage flashcards
  description: 'Set up user accounts, create or manage flashcards, set learning goals, and track performances with user''s heatmap and streak data. Data is available in JSON format only. <br></br>**Note**: All parameters are case-sensitive. Request body parameters are required, but the query parameters are optional. '
  contact:
    name: Dolphin Flashcards
    url: 'https://dolphinflashcards.com/contact'
    email: dolphinflashcards@outlook.com
  license:
    name: MIT
    url: 'https://github.com/jacob-macleod/Dolphin-Flashcard-App/blob/main/LICENSE'
servers:
  - url: 'http://dolphinflashcards.com/api'
    description: baseurl
paths:
...
  /create-xp-goal:
    post:
      summary: Create XP Goal
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                x-examples:
                  Example 1:
                    success: Goal created successfully
                properties:
                  success:
                    type: string
                    example: Goal created successfully
                    description: Provides additional information confirming that the request was successful.
        '400':
          description: A bad request error. Returned when the request body parameters are not in the expected JSON format.
          content:
            application/json:
              schema:
                type: object
                x-examples:
                  Example 1:
                    error: 'Your supplied json keys do not match the expected format. The request should be in the format:  {''jwtToken'': ''string'', ''goalXP'': ''number'', ''endDate'': ''sample-date''}'
                properties:
                  error:
                    type: string
                    example: 'Your supplied json keys do not match the expected format. The request should be in the format: {''jwtToken'': '' '', ''goalXP'': '' '', ''endDate'': '' ''}'
                    description: 'Describes the nature of the error and provides a possible solution. '
              examples:
                Example 1:
                  value:
                    error: Your supplied json keys do not match the expected format.
      operationId: post-create-xp-goal
      x-stoplight:
        id: 9wibu13wmmqj7
      description: 'Create an experience points (XP) goal for the user. XP goals have an `ID`, `type`, `title`, `end date`, `status` (failed, completed, or in progress), `fail date` (if failed), and `data storing` (start date, starting XP, desired XP) properties.'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              x-examples:
                Example 1:
                  jwtToken: '123'
                  goalXP: '100'
                  endDate: 30/07/2024
              required:
                - jwtToken
                - goalXP
                - endDate
              properties:
                jwtToken:
                  type: string
                  example: eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6ImEzMmZkZDRiMTQ2Njc3NzE5YWIyMzcyODYxYmRlZDg5In0.eyJpc3MiOiJodHRwOi8vZG9scGhpbmZsYXNoY2FyZHMuY29tIiwiYXVkIjoiYXBpIiwic3ViIjoidU5RR2lIencxdk9EaExybTVDRWtaQ1ZoTU93MSIsImFjY2Vzc190b2tlbiI6IjRiZTA2NDNmLTFkOTgtNTczYi05N2NkLWNhOThhNjUzNDdkZCIsImFjY2Vzc190b2tlbl9yYXciOiJ0ZXN0IiwiaWF0IjoxNzI2Nzc0MTYzfQ.F-fxFitPQF-gGHyhGPGxE_zj_b8XJybqe4J_iYG66BC4Dqm4LTb8RpcDLyYUhRCcpphihpWGbdp
                  description: Identifies the user who is creating an XP goal.
                goalXP:
                  type: string
                  example: '100'
                  description: The desired amount of experience points (XP) the user aims to achieve with this goal.
                endDate:
                  type: string
                  example: 30/07/2024
                  description: 'The date by which the user intends to achieve the XP goal, in dd/mm/yyyy format.'
            examples:
              Example 1:
                value:
                  jwtToken: '123'
                  goalXP: '100'
                  endDate: 30/07/2024
      tags:
        - Goal
    servers:
      - url: 'http://dolphinflashcards.com/api'
        description: create-xp
  /create-card-goal:
    post:
      summary: Create Card Goal
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                x-examples:
                  Example 1:
                    success: Goal created successfully
                properties:
                  success:
                    type: string
                    description: ' Provides additional information confirming that the request was successful.'
                    example: Goal created successfully
        '400':
          description: A bad request error. Returned when the request body parameters are not in the expected JSON format.
          content:
            application/json:
              schema:
                type: object
                x-examples:
                  Example 1:
                    error: 'Your supplied json keys do not match the expected format. The request should be in the format: {''jwtToken'': '''', ''cardsToRevise'': '''', ''endDate'': ''''}'
                properties:
                  error:
                    type: string
                    description: ' Describes the nature of the error and provides a possible solution.   '
                    example: 'Your supplied json keys do not match the expected format. The request should be in the format: {''jwtToken'': '' '', ''cardsToRevise'': '' '', ''endDate'': '' ''}'
      operationId: post-create-card-goal
      x-stoplight:
        id: 35ujdrtoyfzr9
      description: "Create a card goal for the user. Card goals have the following properties: `ID`, `type` (XP), `title`, `end date`, `status` (failed, completed, or in progress), `fail date` (if failed), and `data storing` (cards revised so far, starting XP, desired cards to revise). A card goal is similar to an XP goal but differs in the data section.\r\n\r\n"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              x-examples:
                Example 1:
                  jwtToken: eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6ImEzMmZkZDRiMTQ2Njc3NzE5YWIyMzcyODYxYmRlZDg5In0.eyJpc3MiOiJodHRwOi8vZG9scGhpbmZsYXNoY2FyZHMuY29tIiwiYXVkIjoiYXBpIiwic3ViIjoidU5RR2lIencxdk9EaExybTVDRWtaQ1ZoTU93MSIsImFjY2Vzc190b2tlbiI6IjRiZTA2NDNmLTFkOTgtNTczYi05N2NkLWNhOThhNjUzNDdkZCIsImFjY2Vzc190b2tlbl9yYXciOiJ0ZXN0IiwiaWF0IjoxNzI2Nzc0MTYzfQ.F-fxFitPQF-gGHyhGPGxE_zj_b8XJybqe4J_iYG66BC4Dqm4LTb8RpcDLyYUhRCcpphihpWGbdp
                  cardsToRevise: '4'
                  endDate: 30/07/2024
              required:
                - jwtToken
                - cardsToRevise
                - endDate
              properties:
                jwtToken:
                  type: string
                  description: Identifies the user who is creating a card goal.
                  example: eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6ImEzMmZkZDRiMTQ2Njc3NzE5YWIyMzcyODYxYmRlZDg5In0.eyJpc3MiOiJodHRwOi8vZG9scGhpbmZsYXNoY2FyZHMuY29tIiwiYXVkIjoiYXBpIiwic3ViIjoidU5RR2lIencxdk9EaExybTVDRWtaQ1ZoTU93MSIsImFjY2Vzc190b2tlbiI6IjRiZTA2NDNmLTFkOTgtNTczYi05N2NkLWNhOThhNjUzNDdkZCIsImFjY2Vzc190b2tlbl9yYXciOiJ0ZXN0IiwiaWF0IjoxNzI2Nzc0MTYzfQ.F-fxFitPQF-gGHyhGPGxE_zj_b8XJybqe4J_iYG66BC4Dqm4LTb8RpcDLyYUhRCcpphihpWGbdp
                cardsToRevise:
                  type: string
                  description: The number of cards the user intends to revise for the current goal.
                  example: '5'
                endDate:
                  type: string
                  description: "The date by which the user intends to achieve the card revision goal, in dd/mm/yyyy format.\t"
                  example: 30/07/2024
      tags:
        - Goal
    servers:
      - url: 'http://dolphinflashcards.com/api'
        description: ''
  /update-goal-status:
    post:
      summary: Update Goal Status
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                x-examples:
                  Example 1:
                    '27730416841422320452490899630401426960300250535420717368523585221911685508588':
                      data:
                        cards_revised_so_far: '0'
                        cards_to_revise: '4'
                      end_date: 30/07/2024
                      fail_date: ''
                      status: in progress
                      title: Revise 4 cards by 07/07/2024
                      type: Card
                    '46882192866017507647804408894345572659406282731228567391554804537877046919642':
                      data:
                        goal_xp: '60'
                        start_date: 15/07/2024
                        starting_xp: '0'
                      end_date: 30/07/2024
                      fail_date: ''
                      status: in progress
                      title: Gain 60 XP by 30/07/2024
                      type: XP
                properties:
                  goalID-1:
                    type: object
                    x-stoplight:
                      id: vp5tii3oejq4d
                    description: 'A unique identifier or key associated with each goal. '
                    properties:
                      data:
                        type: object
                        description: ' An object containing goal-specific metrics, such as `goal_xp`, `start_date`, and `starting_xp` for XP goals or `cards_to_revise` and `cards_revised_so_far` for card goals '
                        properties:
                          cards_revised_so_far:
                            type: string
                            description: 'Associated with card goals, this indicates the number of cards revised so far.'
                            example: '0'
                          cards_to_revise:
                            type: string
                            description: 'Associated with card goals, this specifies the total number of cards to be revised.'
                            example: '4'
                      end_date:
                        type: string
                        example: 30/07/2024
                        description: 'The date by which the current goal must end, in dd/mm/yyyy format.'
                      fail_date:
                        type: string
                        example: '" "'
                        description: 'The date on which the goal''s status changed to ''failed''. This field is only populated if the goal fails; otherwise, it is an empty string.'
                      status:
                        type: string
                        example: in progress
                        description: 'The current status of a goal, which can be either `completed`, `in progress`, or `failed`.'
                      title:
                        type: string
                        example: Revise 4 cards by 15/07/2024
                        description: 'The name assigned to a goal. '
                      type:
                        type: string
                        example: Card
                        description: ' The nature of a goal. Goals can either be of type `Card` or `XP`. '
                  goalID-2:
                    type: object
                    x-stoplight:
                      id: 4f9jwlnkfq031
                    description: 'A unique identifier or key associated with each goal. '
                    properties:
                      data:
                        type: object
                        description: ' An object containing goal-specific metrics, such as `goal_xp`, `start_date`, and `starting_xp` for XP goals or `cards_to_revise` and `cards_revised_so_far` for card goals '
                        properties:
                          goal_xp:
                            type: string
                            description: 'Associated with XP goals, this specifies the target amount of experience points (XP) the user aims to achieve.'
                            example: '60'
                          start_date:
                            type: string
                            description: ' Associated with XP goals, this indicates the date when the XP goal began. '
                            example: 15/07/2024
                          starting_xp:
                            type: string
                            description: 'Associated with XP goals, this indicates the initial amount of experience points (XP) at the beginning of the goal.'
                            example: '0'
                      end_date:
                        type: string
                        description: 'The date by which the current goal must end, in dd/mm/yyyy format.'
                        example: 30/07/2024
                      fail_date:
                        type: string
                        description: 'The date on which the goal''s status changed to ''failed''. This field is only populated if the goal fails; otherwise, it is an empty string.'
                        example: ' " "'
                      status:
                        type: string
                        description: 'The current status of a goal, which can be either `completed`, `in progress`, or `failed`.'
                        example: in progress
                      title:
                        type: string
                        description: 'The name assigned to a goal. '
                        example: Gain 60 XP by 30/07/2024
                      type:
                        type: string
                        description: ' The nature of a goal. Goals can either be of type `Card` or `XP`. '
                        example: XP
        '400':
          description: A bad request error. Returned when the request body parameter is not in the expected JSON format.
          content:
            application/json:
              schema:
                type: object
                x-examples:
                  Example 1:
                    error: 'Your supplied json keys do not match the expected format. The request should be in the format: {''jwtToken'': ''string''}'
                properties:
                  error:
                    type: string
                    example: 'Your supplied json keys do not match the expected format. The request should be in the format: {''jwtToken'': '' ''} '
                    description: Describes the nature of the error and provides a possible solution.
      operationId: post-update-goal-status
      x-stoplight:
        id: hmtbfg51sa572
      description: 'Update the status of a goal. The possible values for the status of a goal are `completed`, `in progress`, or `failed`. Returns an empty object when the jwtToken provided is invalid.'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - jwtToken
              properties:
                jwtToken:
                  type: string
                  x-stoplight:
                    id: z79bhm434n5vx
                  description: An identifier for the user who is updating the goal status.
                  example: eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6ImEzMmZkZDRiMTQ2Njc3NzE5YWIyMzcyODYxYmRlZDg5In0.eyJpc3MiOiJodHRwOi8vZG9scGhpbmZsYXNoY2FyZHMuY29tIiwiYXVkIjoiYXBpIiwic3ViIjoidU5RR2lIencxdk9EaExybTVDRWtaQ1ZoTU93MSIsImFjY2Vzc190b2tlbiI6IjRiZTA2NDNmLTFkOTgtNTczYi05N2NkLWNhOThhNjUzNDdkZCIsImFjY2Vzc190b2tlbl9yYXciOiJ0ZXN0IiwiaWF0IjoxNzI2Nzc0MTYzfQ.F-fxFitPQF-gGHyhGPGxE_zj_b8XJybqe4J_iYG66BC4Dqm4LTb8RpcDLyYUhRCcpphihpWGbdp
      tags:
        - Goal
  /edit-card-goal:
    post:
      summary: Edit Card Goal
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                x-examples:
                  Example 1:
                    success: Goal updated successfully
                properties:
                  success:
                    type: string
                    example: Goal updated successfully
                    description: Provides additional information confirming that the request was successful.
        '400':
          description: A bad request error. Returned when the request body parameters are not in the expected JSON format.
          content:
            application/json:
              schema:
                type: object
                x-examples:
                  Example 1:
                    error: 'Your supplied json keys do not match the expected format. The request should be in the format: {''jwtToken'': '''', ''goalID'': '''', ''newEndDate'': '''', ''newTitle'': '''', ''newCardsToRevise'': ''''}'
                properties:
                  error:
                    type: string
                    example: 'Your supplied json keys do not match the expected format. The request should be in the format: {''jwtToken'': '' '', ''goalID'': '' '', ''newEndDate'': '''', ''newTitle'': '' '', ''newCardsToRevise'': '' ''}'
                    description: Describes the nature of the error and provides a possible solution.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                x-examples:
                  Example 1:
                    error: Goal not found
                properties:
                  error:
                    type: string
                    example: Goal not found
                    description: Describes the nature of the error.
      operationId: post-edit-card-goal
      x-stoplight:
        id: 572gxxxvksi4z
      description: Edit an existing card goal for the user.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              x-examples:
                Example 1:
                  jwtToken: eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6ImEzMmZkZDRiMTQ2Njc3NzE5YWIyMzcyODYxYmRlZDg5In0.eyJpc3MiOiJodHRwOi8vZG9scGhpbmZsYXNoY2FyZHMuY29tIiwiYXVkIjoiYXBpIiwic3ViIjoidU5RR2lIencxdk9EaExybTVDRWtaQ1ZoTU93MSIsImFjY2Vzc190b2tlbiI6IjRiZTA2NDNmLTFkOTgtNTczYi05N2NkLWNhOThhNjUzNDdkZCIsImFjY2Vzc190b2tlbl9yYXciOiJ0ZXN0IiwiaWF0IjoxNzI2Nzc0MTYzfQ.F-fxFitPQF-gGHyhGPGxE_zj_b8XJybqe4J_iYG66BC4Dqm4LTb8RpcDLyYUhRCcpphihpWGbdp
                  goalID: '58748748168480479028433924465493243860761911102946223297681236368808083325563'
                  newEndDate: 30/07/2024
                  newTitle: Updated Goal
                  newCardsToRevise: 8
              required:
                - jwtToken
                - goalID
                - newEndDate
                - newTitle
                - newCardsToRevise
              properties:
                jwtToken:
                  type: string
                  description: 'Stores the user who is editing their card goal.  '
                  example: eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6ImEzMmZkZDRiMTQ2Njc3NzE5YWIyMzcyODYxYmRlZDg5In0.eyJpc3MiOiJodHRwOi8vZG9scGhpbmZsYXNoY2FyZHMuY29tIiwiYXVkIjoiYXBpIiwic3ViIjoidU5RR2lIencxdk9EaExybTVDRWtaQ1ZoTU93MSIsImFjY2Vzc190b2tlbiI6IjRiZTA2NDNmLTFkOTgtNTczYi05N2NkLWNhOThhNjUzNDdkZCIsImFjY2Vzc190b2tlbl9yYXciOiJ0ZXN0IiwiaWF0IjoxNzI2Nzc0MTYzfQ.F-fxFitPQF-gGHyhGPGxE_zj_b8XJybqe4J_iYG66BC4Dqm4LTb8RpcDLyYUhRCcpphihpWGbdp
                goalID:
                  type: string
                  description: A unique identifier or key associated with the card goal being edited.
                  example: '11110748168480479028433924465493243860761911102946223297681236368808083322222'
                newEndDate:
                  type: string
                  description: 'The new end date for the card goal, in `dd/mm/yyyy` format. '
                  example: 30/07/2024
                newTitle:
                  type: string
                  description: The new title assigned to the card goal.
                  example: Updated Goal
                newCardsToRevise:
                  type: string
                  description: 'Specifies the updated number of cards the user intends to revise for this goal. '
                  example: '4'
            examples:
              Example 1:
                value:
                  jwtToken: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6ImEzMmZkZDRiMTQ2Njc3NzE5YWIyMzcyODYxYmRlZDg5In0.eyJpc3MiOiJodHRwOi8vZG9scGhpbmZsYXNoY2FyZHMuY29tIiwiYXVkIjoiYXBpIiwic3ViIjoidU5RR2lIencxdk9EaExybTVDRWtaQ1ZoTU93MSIsImFjY2Vzc190b2tlbiI6IjRiZTA2NDNmLTFkOTgtNTczYi05N2NkLWNhOThhNjUzNDdkZCIsImFjY2Vzc190b2tlbl9yYXciOiJ0ZXN0IiwiaWF0IjoxNzI2Nzc0MTYzfQ.F-fxFitPQF-gGHyhGPGxE_zj_b8XJybqe4J_iYG66BC4Dqm4LTb8RpcDLyYUhRCcpphihpWGbdp'
                  goalID: goal1
                  newEndDate: 30/07/2024
                  newTitle: title1
                  newCardsToRevise: '4'
      tags:
        - Goal
  /edit-xp-goal:
    post:
      summary: Edit XP Goal
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                type: object
                x-examples:
                  Example 1:
                    success: Goal updated successfully
                properties:
                  success:
                    type: string
                    example: Goal updated successfully
                    description: Provides additional information confirming that the request was successful.
        '400':
          description: A bad request error. Returned when the request body parameters are not in the expected JSON format.
          content:
            application/json:
              schema:
                type: object
                x-examples:
                  Example 1:
                    error: 'Your supplied json keys do not match the expected format. The request should be in the format: {''jwtToken'': '''', ''goalID'': '''', ''newEndDate'': ''^(0[1-9]|[12][0-9]|3[01])/(0[1-9]|1[0-2])/\\d{4}$'', ''newTitle'': '''', ''newGoalXP'': ''^\\d+$''}'
                properties:
                  error:
                    type: string
                    example: 'Your supplied json keys do not match the expected format. The request should be in the format: {''jwtToken'': '' '', ''goalID'': '' '', ''newEndDate'': '' '', ''newTitle'': '' '', ''newGoalXP'': '' ''}'
                    description: 'Describes the nature of the error and provides a possible solution. '
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                type: object
                x-examples:
                  Example 1:
                    error: Goal not found
                properties:
                  error:
                    type: string
                    example: Goal not found
                    description: 'Describes the nature of the error. '
      operationId: post-edit-xp-goal
      x-stoplight:
        id: o6rdyybbanhpy
      description: Edit an existing experience points (XP) goal for the user.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              x-examples:
                Example 1:
                  jwtToken: eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6ImEzMmZkZDRiMTQ2Njc3NzE5YWIyMzcyODYxYmRlZDg5In0.eyJpc3MiOiJodHRwOi8vZG9scGhpbmZsYXNoY2FyZHMuY29tIiwiYXVkIjoiYXBpIiwic3ViIjoidU5RR2lIencxdk9EaExybTVDRWtaQ1ZoTU93MSIsImFjY2Vzc190b2tlbiI6IjRiZTA2NDNmLTFkOTgtNTczYi05N2NkLWNhOThhNjUzNDdkZCIsImFjY2Vzc190b2tlbl9yYXciOiJ0ZXN0IiwiaWF0IjoxNzI2Nzc0MTYzfQ.F-fxFitPQF-gGHyhGPGxE_zj_b8XJybqe4J_iYG66BC4Dqm4LTb8RpcDLyYUhRCcpphihpWGbdp
                  goalID: '46882192866017507647804408894345572659406282731228567391554804537877046919642'
                  newEndDate: 15/07/2024
                  newTitle: Updated Title for XP Goal
                  newGoalXP: '80'
              required:
                - jwtToken
                - goalID
                - newEndDate
                - newTitle
                - newGoalXP
              properties:
                jwtToken:
                  type: string
                  description: 'Identifies the user who is editing their XP goal.    '
                goalID:
                  type: string
                  description: A unique identifier or key associated with the XP goal being edited.
                  example: '55335192866017507647804408894345572659406282731228567391554804537877046919111'
                newEndDate:
                  type: string
                  description: ' The new end date for the XP goal, formatted as `dd/mm/yyyy`. '
                  example: 15/07/2024
                newTitle:
                  type: string
                  example: Updated Title for XP Goal
                  description: 'The new title assigned to the XP goal. '
                newGoalXP:
                  type: string
                  example: '80'
                  description: The new desired amount of experience points (XP) the user intends to achieve with this goal.
            examples:
              Example 1:
                value:
                  jwtToken: eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiIsImtpZCI6ImEzMmZkZDRiMTQ2Njc3NzE5YWIyMzcyODYxYmRlZDg5In0.eyJpc3MiOiJodHRwOi8vZG9scGhpbmZsYXNoY2FyZHMuY29tIiwiYXVkIjoiYXBpIiwic3ViIjoidU5RR2lIencxdk9EaExybTVDRWtaQ1ZoTU93MSIsImFjY2Vzc190b2tlbiI6IjRiZTA2NDNmLTFkOTgtNTczYi05N2NkLWNhOThhNjUzNDdkZCIsImFjY2Vzc190b2tlbl9yYXciOiJ0ZXN0IiwiaWF0IjoxNzI2Nzc0MTYzfQ.F-fxFitPQF-gGHyhGPGxE_zj_b8XJybqe4J_iYG66BC4Dqm4LTb8RpcDLyYUhRCcpphihpWGbdp
                  goalID: goal1
                  newEndDate: 15/07/2024
                  newTitle: title1
                  newGoalXP: '40'
      tags:
        - Goal
...
components:
...
  securitySchemes:
    apiKey_1:
      name: apiKey
      type: apiKey
      in: query
...